Displayed times are time zone aware (#11)
A time zone is stored in the database for each session, this can be
auto-detected if the session has GPS information, plus the user can manually
edit it.  All displayed dates and times in the session inspector display the
local time at the specified time zone.

* .gitmodules: new file, add required packages as submodules

* .travis.yml, etc/scripts/azure-common.yml: setup package catalog before the
  build.  Also fetch newer versions of the test data.

* build.rkt (main): check that the required modules are installed before
  running a build.

* run.rkt: call `check-missing-modules` to verify that tzinfo and tzgeolookup
  are installed.

* docs/README.md: updated with instructions about the extra setup related to
  package dependencies.

* etc/scripts/fetch.sh: add option do download unencrypted files, also changed
  encryption algorithm, as the old one was deprecatd.

* etc/scripts/install-racket.sh: add support for installing CS Racket
  variants, e.g. when version is 7.4-cs.

* etc/scripts/setup-catalog.sh: new file, sets up the package catalog from the
  git submodules.

* rkt/al-widgets.rkt (interval-choice% add-time-zone): add a time-zone field
  to every interval. All interval types how have a time-zone field inherited
  from the session.  (*run-lap-fields*, *bike-lap-fields*, *swim-lap-fields*)
  (*swim-mini-lap-fields*): "Time of day" field will display local time at the
  specified time zone.  (time-zone-selector%): new widget for choosing a time
  zone

* rkt/database.rkt (fetch-session-stmt): fetch the time zone of the session.

* rkt/dbapp.rkt (upgrade-patches): register patch 32 for adding time zones to
  the database.

* rkt/fmt-util-ut.rkt: new file, contains time-zone aware date and time
  formatting utilities, moved here from fmt-util.rkt

* rkt/import.rkt (update-timezone-for-new-sessions): new defun, also updated
  other pos-import functions to receive a list of the imported sessions,
  rather than queriing it every time.

* rkt/toplevel.rkt (interactive-update-time-zones): new defun, updates time
  zones for all sessions in the database.

* rkt/utilities.rkt (check-missing-modules): new macro, checks for missing
  modules and reports an error.

* rkt/view-activities.rkt (view-activities% activity-list-display-columns*):
  "Start Time" field displays local time depending on time zone.

* rkt/view-last-import.rkt (*display-columns*): "Start Time" field displays
  local time depending on time zone.

* rkt/dialogs/edit-lap-swim.rkt (lap-swim-edit% setup): fetch the time zone
  and display a timestamp in the local time at the specified time zone.

* rkt/dialogs/edit-session-summary.rkt (edit-session-summary-dialog%): allow
the user to edit the time zone of a session.

* rkt/dialogs/edit-session-tss.rkt (edit-session-tss-dialog%): fetch the time
  zone and display a timestamp in the local time at the specified time zone.

* rkt/fit-file/activity-util.rkt (session-time-zone): new defun

* rkt/session-inspector/view-session.rkt (session-header%): display a
  time-zone aware session start time and allow editing the time zone when in
  edit mode.

* rkt/widgets/date-input-field.rkt (date-input-field%): allow displaying local
  or UTC time and allow a time-zone based time conversion.

* rkt/widgets/time-of-day-input-field.rkt (time-of-day-input-field%
  set-time-of-day-value): allow specifying the time zone along with the UTC
  timestamp.

* sql/db-schema.sql (E_TIME_ZONE): new database table holding all time zone
  names.  (A_SESSION time_zone_id): new field referencing an entry into
  `E_TIME_ZONE`

* sql/migrations/p32-time-zone.sql: database migration script to add the
  E_TIME_ZONE table and A_SESSION.time_zone_id fields to an existing database.

* test/test-util.rkt (check-time-zone): check that the timezone stored in the
  database is correct.  (db-import-activity-from-file/check): updated to also
  call `check-time-zone`

+++ etc/scripts/fetch.sh
@@ -18,19 +18,22 @@
 
 set -e
 script_name=${0##**/}
+decrypt=0
 
-if [ -z $TESTDATAPW ]; then
-    echo "$script_name: missing TESTDATAPW env var, will not download anything"
-    exit 0
-fi
-
-while getopts :C: OPT; do
+while getopts :C:d OPT; do
     case $OPT in
         C|+C)
             output_dir="$OPTARG"
             ;;
+        d|+d)
+            if [ -z $TESTDATAPW ]; then
+                echo "$script_name: missing TESTDATAPW env var, will not download anything"
+                exit 0    # must return 0, as we use this from build pipelines
+            fi
+            decrypt=1
+            ;;
         *)
-            echo "usage: ${0##*/} [+-C output_dir} [--] file_id"
+            echo "usage: $script_name [+-C output_dir] [-d] [--] file_id"
             exit 2
     esac
 done
@@ -52,7 +55,6 @@ trap 'rm --force -- $gcookies' INT TERM HUP EXIT
 curl --silent \
      --cookie-jar $gcookies \
      "$gurl?export=download&id=$file_id" > /dev/null
-
 code="$(awk '/_warning_/ {print $NF}' $gcookies)"
 ofile1=$(mktemp ${TMP:-/tmp}/$script_name.XXXXXXXXXX)
 
@@ -64,11 +66,17 @@ curl --location \
      --output $ofile1 \
      "$gurl?export=download&confirm=$code&id=$file_id"
 
-ofile2=$(mktemp ${TMP:-/tmp}/$script_name.XXXXXXXXXX)
-trap 'rm --force -- $ofile2' INT TERM HUP EXIT
-openssl aes-256-cbc -md md5 -d -k $TESTDATAPW -in $ofile1 -out $ofile2
-# We run in the root of the repo, but output test databases into the test
-# folder
+if ((decrypt == 1)); then
+    ofile2=$(mktemp ${TMP:-/tmp}/$script_name.XXXXXXXXXX)
+    trap 'rm --force -- $ofile2' INT TERM HUP EXIT
+    openssl enc -d -aes-256-ctr -md sha256 \
+            -pass pass:$TESTDATAPW -in $ofile1 -out $ofile2
+else
+    ofile2=$ofile1
+fi
+
+mkdir --parent "${output_dir:-.}" > /dev/null 2>&1
 tar xvzf $ofile2 -C "${output_dir:-.}"
+
 # Clean up after ourselves
 rm --force -- $gcookies $ofile1 $ofile2
